test(tmpfs): add unit tests for parsing tmpfs size options
