+++ kernel/src/fs/syscall.rs
@@ -329,78 +329,6 @@ fn parse_overlay_options(data: &str) -> Result<(Option<String>, Vec<String>), ()
     Ok((upperdir, lowerdirs))
 }
 
-// Helper function to parse tmpfs size option
-fn parse_tmpfs_size(data: &str) -> Result<usize, ()> {
-    for option in data.split(',') {
-        if let Some(size_str) = option.strip_prefix("size=") {
-            // Parse size with suffix (K, M, G)
-            let size_str = size_str.trim();
-            if size_str.is_empty() {
-                continue;
-            }
-            
-            let (number_part, multiplier) = if size_str.ends_with('K') || size_str.ends_with('k') {
-                (&size_str[..size_str.len()-1], 1024)
-            } else if size_str.ends_with('M') || size_str.ends_with('m') {
-                (&size_str[..size_str.len()-1], 1024 * 1024)
-            } else if size_str.ends_with('G') || size_str.ends_with('g') {
-                (&size_str[..size_str.len()-1], 1024 * 1024 * 1024)
-            } else {
-                (size_str, 1)
-            };
-            
-            if let Ok(number) = number_part.parse::<usize>() {
-                return Ok(number * multiplier);
-            }
-        }
-    }
-    Err(())
-}
-
-// Helper function to parse memory range (start,end)
-fn parse_memory_range(data: &str) -> Result<(usize, usize), ()> {
-    let parts: Vec<&str> = data.split(',').collect();
-    if parts.len() != 2 {
-        return Err(());
-    }
-    
-    let start = if parts[0].starts_with("0x") {
-        usize::from_str_radix(&parts[0][2..], 16).map_err(|_| ())?
-    } else {
-        parts[0].parse().map_err(|_| ())?
-    };
-    
-    let end = if parts[1].starts_with("0x") {
-        usize::from_str_radix(&parts[1][2..], 16).map_err(|_| ())?
-    } else {
-        parts[1].parse().map_err(|_| ())?
-    };
-    
-    Ok((start, end))
-}
-
-// Helper function to create and mount tmpfs
-fn create_tmpfs_and_mount(_vfs: &VfsManager, _mount_point: &str, _memory_limit: usize) -> Result<(), super::FileSystemError> {
-    // This would typically create tmpfs params and use create_and_register_fs_with_params
-    // For now, simplified implementation
-    // TODO: Implement proper tmpfs parameter handling
-    Err(super::FileSystemError {
-        kind: super::FileSystemErrorKind::NotFound,
-        message: "Tmpfs creation not fully implemented".to_string(),
-    })
-}
-
-// Helper function to create and mount block filesystem
-fn create_block_fs_and_mount(_vfs: &VfsManager, _fstype: &str, _device_path: &str, _mount_point: &str) -> Result<(), super::FileSystemError> {
-    // This would typically open the block device and create filesystem
-    // For now, simplified implementation
-    // TODO: Implement proper block device handling
-    Err(super::FileSystemError {
-        kind: super::FileSystemErrorKind::NotFound,
-        message: "Block device filesystem creation not fully implemented".to_string(),
-    })
-}
-
 /// Create a filesystem using the driver and mount it
 /// 
 /// This function uses the new driver-based approach where option parsing
@@ -427,213 +355,4 @@ fn create_filesystem_and_mount(
     vfs.mount(fs_id, target)?;
     
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test_case]
-    fn test_parse_overlay_options_basic() {
-        // Test basic overlay options with single lowerdir
-        let result = parse_overlay_options("lowerdir=/lower");
-        assert!(result.is_ok());
-        let (upperdir, lowerdirs) = result.unwrap();
-        assert!(upperdir.is_none());
-        assert_eq!(lowerdirs.len(), 1);
-        assert_eq!(lowerdirs[0], "/lower");
-    }
-
-    #[test_case]
-    fn test_parse_overlay_options_with_upper() {
-        // Test overlay options with upperdir and lowerdir
-        let result = parse_overlay_options("upperdir=/upper, lowerdir=/lower");
-        assert!(result.is_ok());
-        let (upperdir, lowerdirs) = result.unwrap();
-        assert!(upperdir.is_some());
-        assert_eq!(upperdir.unwrap(), "/upper");
-        assert_eq!(lowerdirs.len(), 1);
-        assert_eq!(lowerdirs[0], "/lower");
-    }
-
-    #[test_case]
-    fn test_parse_overlay_options_multiple_lower() {
-        // Test overlay options with multiple lowerdirs
-        let result = parse_overlay_options("lowerdir=/lower1:/lower2:/lower3");
-        assert!(result.is_ok());
-        let (upperdir, lowerdirs) = result.unwrap();
-        assert!(upperdir.is_none());
-        assert_eq!(lowerdirs.len(), 3);
-        assert_eq!(lowerdirs[0], "/lower1");
-        assert_eq!(lowerdirs[1], "/lower2");
-        assert_eq!(lowerdirs[2], "/lower3");
-    }
-
-    #[test_case]
-    fn test_parse_overlay_options_complex() {
-        // Test complex overlay options
-        let result = parse_overlay_options("upperdir=/upper,lowerdir=/lower1:/lower2,workdir=/work");
-        assert!(result.is_ok());
-        let (upperdir, lowerdirs) = result.unwrap();
-        assert!(upperdir.is_some());
-        assert_eq!(upperdir.unwrap(), "/upper");
-        assert_eq!(lowerdirs.len(), 2);
-        assert_eq!(lowerdirs[0], "/lower1");
-        assert_eq!(lowerdirs[1], "/lower2");
-    }
-
-    #[test_case]
-    fn test_parse_overlay_options_no_lowerdir() {
-        // Test overlay options without lowerdir (should fail)
-        let result = parse_overlay_options("upperdir=/upper");
-        assert!(result.is_err());
-    }
-
-    #[test_case]
-    fn test_parse_overlay_options_empty() {
-        // Test empty options (should fail)
-        let result = parse_overlay_options("");
-        assert!(result.is_err());
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_bytes() {
-        // Test parsing size in bytes
-        let result = parse_tmpfs_size("size=1048576");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 1048576);
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_kilobytes() {
-        // Test parsing size in kilobytes
-        let result = parse_tmpfs_size("size=10K");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 10 * 1024);
-
-        let result = parse_tmpfs_size("size=5k");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 5 * 1024);
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_megabytes() {
-        // Test parsing size in megabytes
-        let result = parse_tmpfs_size("size=64M");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 64 * 1024 * 1024);
-
-        let result = parse_tmpfs_size("size=128m");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 128 * 1024 * 1024);
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_gigabytes() {
-        // Test parsing size in gigabytes
-        let result = parse_tmpfs_size("size=2G");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 2 * 1024 * 1024 * 1024);
-
-        let result = parse_tmpfs_size("size=1g");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 1 * 1024 * 1024 * 1024);
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_multiple_options() {
-        // Test parsing size with multiple options
-        let result = parse_tmpfs_size("nodev,size=32M,noexec");
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), 32 * 1024 * 1024);
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_no_size() {
-        // Test parsing without size option (should fail)
-        let result = parse_tmpfs_size("nodev,noexec");
-        assert!(result.is_err());
-    }
-
-    #[test_case]
-    fn test_parse_tmpfs_size_invalid() {
-        // Test parsing invalid size
-        let result = parse_tmpfs_size("size=invalid");
-        assert!(result.is_err());
-
-        let result = parse_tmpfs_size("size=");
-        assert!(result.is_err());
-    }
-
-    #[test_case]
-    fn test_parse_memory_range_decimal() {
-        // Test parsing memory range in decimal
-        let result = parse_memory_range("134217728,268435456");
-        assert!(result.is_ok());
-        let (start, end) = result.unwrap();
-        assert_eq!(start, 134217728);
-        assert_eq!(end, 268435456);
-    }
-
-    #[test_case]
-    fn test_parse_memory_range_hex() {
-        // Test parsing memory range in hexadecimal
-        let result = parse_memory_range("0x80000000,0x81000000");
-        assert!(result.is_ok());
-        let (start, end) = result.unwrap();
-        assert_eq!(start, 0x80000000);
-        assert_eq!(end, 0x81000000);
-    }
-
-    #[test_case]
-    fn test_parse_memory_range_mixed() {
-        // Test parsing memory range with mixed formats
-        let result = parse_memory_range("0x80000000,268435456");
-        assert!(result.is_ok());
-        let (start, end) = result.unwrap();
-        assert_eq!(start, 0x80000000);
-        assert_eq!(end, 268435456);
-    }
-
-    #[test_case]
-    fn test_parse_memory_range_invalid_format() {
-        // Test parsing invalid memory range formats
-        let result = parse_memory_range("0x80000000");
-        assert!(result.is_err());
-
-        let result = parse_memory_range("0x80000000,0x81000000,extra");
-        assert!(result.is_err());
-
-        let result = parse_memory_range("");
-        assert!(result.is_err());
-    }
-
-    #[test_case]
-    fn test_parse_memory_range_invalid_numbers() {
-        // Test parsing invalid numbers
-        let result = parse_memory_range("invalid,0x81000000");
-        assert!(result.is_err());
-
-        let result = parse_memory_range("0x80000000,invalid");
-        assert!(result.is_err());
-
-        let result = parse_memory_range("0xinvalid,0x81000000");
-        assert!(result.is_err());
-    }
-
-    #[test_case]
-    fn test_parse_memory_range_boundary_cases() {
-        // Test boundary cases
-        let result = parse_memory_range("0,1");
-        assert!(result.is_ok());
-        let (start, end) = result.unwrap();
-        assert_eq!(start, 0);
-        assert_eq!(end, 1);
-
-        let result = parse_memory_range("0x0,0xffffffff");
-        assert!(result.is_ok());
-        let (start, end) = result.unwrap();
-        assert_eq!(start, 0);
-        assert_eq!(end, 0xffffffff);
-    }
 }
